About Me!
I study Computational Design.
Programming to me is like learning an exotic language. It's hard to learn at first but I know it will be super rewarding once I get to solve problems with it.
I want to be able to mine, manipulate, analyse and visualize data. I also want to get into web development and program arduino kits, to say the least.